Changes to the tests to make them run cleanly under 3.x
In general the changes are not large, but they did spot one
potential issue with round(). Since the rounding strategy
changed between 2.x and 3.x there will be different results
between the two, and it is not clear to me that one is
actually superior to the other. As a result I've simply
changed the test to accept both results, but this should
be investigated more thoroughly if time permits.